Sleep the job if not enough time has elapsed
